requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
requests>=2.31
python-dotenv>=1.0